requirements.txt
pyautogui
Pillow
pytesseract
PySide6
keyboard
